pyproject.toml
[tool.poetry]
name = "haystack-ai-playground"
version = "0.1.0"
description = ""
authors = ["vikyw89 <vikyw89@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
haystack-ai = "^2.0.0b2"
python-dotenv = "^1.0.0"
boilerpy3 = "^1.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
